[Bronze II] Title: 캔디 구매, Time: 120 ms, Memory: 14184 KB -BaekjoonHub

+++ 백준/Bronze/2909. 캔디 구매/캔디 구매.java
@@ -12,7 +12,7 @@ public class Main {
         int b = (int) Math.pow(10, Integer.parseInt(st.nextToken()));
         double v = a / (double) b;
 
-        long x = (long) (Math.round(v) * b);
+        int x = (int) (Math.round(v) * b);
         System.out.println(x);
     }
 }
